move RestInput in its own module
