+++ changelog/6199.misc.md
@@ -1,0 +1,1 @@
+Move `RestInput` from `rasa.core.channels.channel` to `rasa.core.channels.rest`.
